+++ src/Hedwig/ClientApp/src/containers/Reports/ReportDetail/RevenueView/UtilizationTable.tsx
@@ -2,68 +2,65 @@ import React from 'react';
 import Table, { TableProps } from '../../../../components/Table/Table';
 import {
 	CdcReport,
-	Enrollment,
-	Age,
-	FundingTime,
-	FundingSpace,
-	FundingSource,
 } from '../../../../generated';
-import idx from 'idx';
-import moment from 'moment';
 import {
 	prettyAge,
 	prettyFundingTime,
-	fundingSpaceSorter,
 	getReportingPeriodWeeks,
-	getFundingSpaces,
 } from '../../../../utils/models';
 import currencyFormatter from '../../../../utils/currencyFormatter';
 import cx from 'classnames';
-import { DeepNonUndefineableArray } from '../../../../utils/types';
 import {
-	countFundedEnrollments,
-	calculateRate,
 	makePrefixerFunc,
 	ReimbursementRateLine,
-	productOfUnknowns,
 } from '../../../../utils/utilizationTable';
 import FormContext, { useGenericContext } from '../../../../components/Form_New/FormContext';
-import { getSplitUtilizations } from '../../../../utils/models/fundingTimeSplitUtilization';
+import { RevenueDataRow, getRevenueDataRows } from '../utils/getRevenueDataRows';
 
-interface UtilizationTableRow {
-	key: string;
-	ageGroup?: Age;
-	fundingTime?: FundingTime;
-	count: number;
-	capacity: number;
-	ftRate?: number;
-	ptRate?: number;
-	total: number;
-	balance: number;
+type UtilizationTableRow = RevenueDataRow & {
 	showFundingTime?: boolean;
-	weeksSplit?: {
-		partWeeks: number | undefined;
-		fullWeeks: number | undefined;
-	};
+	// Used for numeral alignment formatting
 	maxes: {
-		// These are needed to format the numbers for alignment
 		total: number;
 		rate: number;
 		balance: number;
 	};
 }
 
-// Not functional component because ts had a problem with the empty fragments
 export default function UtilizationTable() {
 	const { data: report } = useGenericContext<CdcReport>(FormContext);
 
-	const rows = getUtilizationTableRows(report);
-	const totalRow = rows[rows.length - 1];
+	// Transform RevenueDataRow into UtilizationTableRow
+	let rows: UtilizationTableRow[] = getRevenueDataRows(report).map((row) => ({
+		...row,
+		showFundingTime: (report.organization?.fundingSpaces || []).filter((fs) => fs.ageGroup === row.ageGroup).length > 1,
+		maxes: {
+			total: row.total,
+			balance: row.balance,
+			rate: Math.max(row.ptRate || 0, row.ftRate || 0)
+		}
+	}));
 
+	// And do reduce transformation for total row
+	let totalRow = rows[rows.length - 1];
+	totalRow = rows.reduce(
+		(total, row) => ({
+			...total,
+			maxes: {
+				total: Math.max(total.total, row.total),
+				rate: Math.max(total.maxes.rate, row.ftRate || 0 , row.ptRate || 0),
+				balance: Math.max(Math.abs(total.balance), Math.abs(row.balance))
+			}
+		}),
+		totalRow
+	)
+
+	// Get prefixers
 	const reimbursementPrefixer = makePrefixerFunc(totalRow.maxes.rate);
 	const totalPrefixer = makePrefixerFunc(totalRow.maxes.total);
 	const balancePrefixer = makePrefixerFunc(totalRow.maxes.balance);
 
+	// Instantiate table props
 	const tableProps: TableProps<UtilizationTableRow> = {
 		id: 'utilization-table',
 		data: rows,
@@ -172,128 +169,3 @@ export default function UtilizationTable() {
 		</>
 	);
 }
-
-export const getUtilizationTableRows = (report: CdcReport) => {
-	// TODO: if the space lives on the organization but the rates live on the site,
-	// we need to reconsider how we handle multi site org rate calculation
-	const site = idx(report, (_) => _.organization.sites[0]);
-	const fundingSpaces = idx(report, (_) => _.organization.fundingSpaces);
-
-	if (!site || !fundingSpaces) return [];
-
-	const weeksInPeriod = getReportingPeriodWeeks(report.reportingPeriod);
-	const splitTimeFundingSpaces = getFundingSpaces(
-		report.organization && report.organization.fundingSpaces,
-		{
-			time: FundingTime.Split,
-		}
-	);
-	const timeSplitUtilizations = getSplitUtilizations(report, splitTimeFundingSpaces, weeksInPeriod);
-
-	const enrollments = (idx(report, (_) => _.enrollments) || []) as Enrollment[];
-
-	const cdcFundingSpaces = (fundingSpaces as DeepNonUndefineableArray<FundingSpace>)
-		.sort(fundingSpaceSorter)
-		.filter((fundingSpace) => fundingSpace.source === FundingSource.CDC);
-
-	let rows: UtilizationTableRow[] = cdcFundingSpaces.map((space) => {
-		const capacity = space.capacity;
-		const ageGroup = space.ageGroup;
-		const fundingTime = space.time;
-		const count = countFundedEnrollments(enrollments, space.id);
-		const ptRate = calculateRate(
-			report.accredited,
-			site.titleI,
-			site.region,
-			ageGroup,
-			FundingTime.Part
-		);
-		const ftRate = calculateRate(
-			report.accredited,
-			site.titleI,
-			site.region,
-			ageGroup,
-			FundingTime.Full
-		);
-
-		let fullWeeks: number | undefined, partWeeks: number | undefined;
-		if (fundingTime === FundingTime.Split) {
-			const thisUtilization = (timeSplitUtilizations || []).find((u) => u.reportId === report.id);
-			fullWeeks = 0;
-			partWeeks = 0;
-			if (thisUtilization) {
-				fullWeeks = thisUtilization.fullTimeWeeksUsed || 0;
-				partWeeks = thisUtilization.partTimeWeeksUsed || 0;
-			}
-		} else if (fundingTime === FundingTime.Full) {
-			fullWeeks = weeksInPeriod;
-		} else if (fundingTime === FundingTime.Part) {
-			partWeeks = weeksInPeriod;
-		}
-
-		const cappedCount = Math.min(count, capacity);
-		const total =
-			productOfUnknowns([cappedCount, fullWeeks, ftRate]) +
-			productOfUnknowns([cappedCount, partWeeks, ptRate]);
-		const paid =
-			productOfUnknowns([capacity, fullWeeks, ftRate]) +
-			productOfUnknowns([capacity, partWeeks, ptRate]);
-
-		const balance = total - paid;
-
-		return {
-			key: `${ageGroup}-${fundingTime}`,
-			ageGroup,
-			fundingTime,
-			count,
-			capacity,
-			ftRate,
-			ptRate,
-			total,
-			balance,
-			showFundingTime: cdcFundingSpaces.filter((fs) => fs.ageGroup === ageGroup).length > 1,
-			weeksSplit: {
-				partWeeks: partWeeks,
-				fullWeeks: fullWeeks,
-			},
-			maxes: {
-				total,
-				rate: Math.max(ptRate, ftRate),
-				balance,
-			},
-		};
-	});
-
-	const totalRow = rows.reduce(
-		(total, row) => {
-			return {
-				...total,
-				count: total.count + row.count,
-				capacity: total.capacity + row.capacity,
-				total: total.total + row.total,
-				balance: total.balance + row.balance,
-				maxes: {
-					total: Math.max(total.total, row.total),
-					rate: Math.max(total.maxes.rate, row.ftRate || 0, row.ptRate || 0),
-					balance: Math.max(Math.abs(total.balance), Math.abs(row.balance)),
-				},
-			};
-		},
-		{
-			key: 'total',
-			count: 0,
-			capacity: 0,
-			total: 0,
-			balance: 0,
-			maxes: {
-				total: 0,
-				rate: 0,
-				balance: 0,
-			},
-		}
-	);
-
-	rows.push(totalRow);
-
-	return rows;
-};
